Agregar manejo de ingresos en pedidos: actualizar y eliminar órdenes con referencia a ingresos, y ajustar la interfaz para incluir el valor total.

+++ src/lib/server/revenue.ts
@@ -13,6 +13,8 @@ export async function createRevenue (value: number, reason?: string) {
     }
 
     await db.insert(table.revenue).values(revenue).execute()
+
+    return revenueId;
 }
 
 export async function getRevenues (page: number = 1, limit: number = 10) {
@@ -30,6 +32,8 @@ export async function getRevenues (page: number = 1, limit: number = 10) {
 }
 
 export async function updateRevenue (id: string, value?: number, reason?: string) {
+    const [order] = await db.select().from(table.order).where(eq(table.order.revenueId, id)).execute();
+    if (order) { return }
     if (typeof value !== 'undefined' && typeof reason !== 'undefined') {
         await db.update(table.revenue).set({value, reason}).where(eq(table.revenue.id, id)).execute();
     } else if (typeof value !== 'undefined') {
@@ -40,6 +44,8 @@ export async function updateRevenue (id: string, value?: number, reason?: string
 }
 
 export async function deleteRevenue (id: string) {
+    const [order] = await db.select().from(table.order).where(eq(table.order.revenueId, id)).execute();
+    if (order) { return }
     await db.delete(table.revenue).where(eq(table.revenue.id, id));
 }
